fix: use the right schema for checking compatibility
This patch fixes how we check schema compatibility for new streams/tables.
Previously, we had our own code for generating an avro schema that was just
used to generate the schema used to check for compatibility. The schema we
actually write to the registry is generated by the avro serializer and
avro converter. Now we use this schema when checking compatibility as well.
