+++ ksql-engine/src/test/java/io/confluent/ksql/util/AvroUtilTest.java
@@ -37,11 +37,13 @@ import io.confluent.ksql.schema.ksql.LogicalSchema.Builder;
 import io.confluent.ksql.schema.ksql.PhysicalSchema;
 import io.confluent.ksql.schema.ksql.SchemaConverters;
 import io.confluent.ksql.schema.ksql.SchemaConverters.ConnectToSqlTypeConverter;
+import io.confluent.ksql.schema.ksql.types.SqlTypes;
 import io.confluent.ksql.serde.Format;
 import io.confluent.ksql.serde.FormatInfo;
 import io.confluent.ksql.serde.KeyFormat;
 import io.confluent.ksql.serde.SerdeOption;
 import io.confluent.ksql.serde.ValueFormat;
+import io.confluent.ksql.serde.avro.AvroSchemas;
 import io.confluent.ksql.serde.connect.ConnectSchemaTranslator;
 import java.io.IOException;
 import java.util.Collections;
@@ -57,7 +59,7 @@ import org.mockito.junit.MockitoJUnitRunner;
 @RunWith(MockitoJUnitRunner.class)
 public class AvroUtilTest {
 
-  private static final SourceName STREAM_NAME = SourceName.of("some-stream");
+  private static final SourceName STREAM_NAME = SourceName.of("some_stream");
 
   private static final String AVRO_SCHEMA_STRING = "{"
       + "\"namespace\": \"some.namespace\","
@@ -90,6 +92,11 @@ public class AvroUtilTest {
   private static final LogicalSchema SINGLE_FIELD_SCHEMA =
       toKsqlSchema(SINGLE_FIELD_AVRO_SCHEMA_STRING);
 
+  private static final LogicalSchema SCHEMA_WITH_MAPS = LogicalSchema.builder()
+      .valueColumn(ColumnName.of("notmap"), SqlTypes.BIGINT)
+      .valueColumn(ColumnName.of("mapcol"), SqlTypes.map(SqlTypes.INTEGER))
+      .build();
+
   private static final KsqlTopic RESULT_TOPIC = new KsqlTopic(
       "actual-name",
       KeyFormat.nonWindowed(FormatInfo.of(Format.KAFKA)),
@@ -103,6 +110,8 @@ public class AvroUtilTest {
   private SchemaRegistryClient srClient;
   @Mock
   private CreateSourceCommand ddlCommand;
+
+  private final KsqlConfig ksqlConfig = new KsqlConfig(Collections.emptyMap());
 
   @Before
   public void setUp() {
@@ -118,7 +127,7 @@ public class AvroUtilTest {
     when(srClient.testCompatibility(anyString(), any())).thenReturn(true);
 
     // When:
-    AvroUtil.throwOnInvalidSchemaEvolution(STATEMENT_TEXT, ddlCommand, srClient);
+    AvroUtil.throwOnInvalidSchemaEvolution(STATEMENT_TEXT, ddlCommand, srClient, ksqlConfig);
 
     // Then:
     verify(srClient).testCompatibility(eq(RESULT_TOPIC.getKafkaTopicName() + "-value"), any());
@@ -129,12 +138,31 @@ public class AvroUtilTest {
     // Given:
     final PhysicalSchema schema = PhysicalSchema.from(MUTLI_FIELD_SCHEMA, SerdeOption.none());
 
-    final org.apache.avro.Schema expectedAvroSchema = SchemaUtil
-        .buildAvroSchema(schema.valueSchema(), STREAM_NAME.name());
+    final org.apache.avro.Schema expectedAvroSchema = AvroSchemas
+        .getAvroSchema(schema.valueSchema(), STREAM_NAME.name(), ksqlConfig);
     when(srClient.testCompatibility(anyString(), any())).thenReturn(true);
 
     // When:
-    AvroUtil.throwOnInvalidSchemaEvolution(STATEMENT_TEXT, ddlCommand, srClient);
+    AvroUtil.throwOnInvalidSchemaEvolution(STATEMENT_TEXT, ddlCommand, srClient, ksqlConfig);
+
+    // Then:
+    verify(srClient).testCompatibility(any(), eq(expectedAvroSchema));
+  }
+
+  @Test
+  public void shouldValidateSchemaWithMaps() throws Exception {
+    // Given:
+    when(ddlCommand.getSchema()).thenReturn(SCHEMA_WITH_MAPS);
+    final PhysicalSchema schema = PhysicalSchema
+        .from(SCHEMA_WITH_MAPS, SerdeOption.none());
+
+    when(srClient.testCompatibility(anyString(), any())).thenReturn(true);
+
+    final org.apache.avro.Schema expectedAvroSchema = AvroSchemas
+        .getAvroSchema(schema.valueSchema(), STREAM_NAME.name(), ksqlConfig);
+
+    // When:
+    AvroUtil.throwOnInvalidSchemaEvolution(STATEMENT_TEXT, ddlCommand, srClient, ksqlConfig);
 
     // Then:
     verify(srClient).testCompatibility(any(), eq(expectedAvroSchema));
@@ -149,11 +177,11 @@ public class AvroUtilTest {
 
     when(srClient.testCompatibility(anyString(), any())).thenReturn(true);
 
-    final org.apache.avro.Schema expectedAvroSchema = SchemaUtil
-        .buildAvroSchema(schema.valueSchema(), STREAM_NAME.name());
+    final org.apache.avro.Schema expectedAvroSchema = AvroSchemas
+        .getAvroSchema(schema.valueSchema(), STREAM_NAME.name(), ksqlConfig);
 
     // When:
-    AvroUtil.throwOnInvalidSchemaEvolution(STATEMENT_TEXT, ddlCommand, srClient);
+    AvroUtil.throwOnInvalidSchemaEvolution(STATEMENT_TEXT, ddlCommand, srClient, ksqlConfig);
 
     // Then:
     verify(srClient).testCompatibility(any(), eq(expectedAvroSchema));
@@ -170,11 +198,11 @@ public class AvroUtilTest {
 
     when(srClient.testCompatibility(anyString(), any())).thenReturn(true);
 
-    final org.apache.avro.Schema expectedAvroSchema = SchemaUtil
-        .buildAvroSchema(schema.valueSchema(), STREAM_NAME.name());
+    final org.apache.avro.Schema expectedAvroSchema = AvroSchemas
+        .getAvroSchema(schema.valueSchema(), STREAM_NAME.name(), ksqlConfig);
 
     // When:
-    AvroUtil.throwOnInvalidSchemaEvolution(STATEMENT_TEXT, ddlCommand, srClient);
+    AvroUtil.throwOnInvalidSchemaEvolution(STATEMENT_TEXT, ddlCommand, srClient, ksqlConfig);
 
     // Then:
     verify(srClient).testCompatibility(any(), eq(expectedAvroSchema));
@@ -186,7 +214,7 @@ public class AvroUtilTest {
     when(srClient.testCompatibility(any(), any())).thenReturn(true);
 
     // When:
-    AvroUtil.throwOnInvalidSchemaEvolution(STATEMENT_TEXT, ddlCommand, srClient);
+    AvroUtil.throwOnInvalidSchemaEvolution(STATEMENT_TEXT, ddlCommand, srClient, ksqlConfig);
   }
 
   @Test
@@ -198,7 +226,7 @@ public class AvroUtilTest {
     expectedException.expectMessage("Cannot register avro schema for actual-name as the schema is incompatible with the current schema version registered for the topic");
 
     // When:
-    AvroUtil.throwOnInvalidSchemaEvolution(STATEMENT_TEXT, ddlCommand, srClient);
+    AvroUtil.throwOnInvalidSchemaEvolution(STATEMENT_TEXT, ddlCommand, srClient, ksqlConfig);
   }
 
   @Test
@@ -208,7 +236,7 @@ public class AvroUtilTest {
         .thenThrow(new RestClientException("Unknown subject", 404, 40401));
 
     // When:
-    AvroUtil.throwOnInvalidSchemaEvolution(STATEMENT_TEXT, ddlCommand, srClient);
+    AvroUtil.throwOnInvalidSchemaEvolution(STATEMENT_TEXT, ddlCommand, srClient, ksqlConfig);
   }
 
   @Test
@@ -227,7 +255,7 @@ public class AvroUtilTest {
     )));
 
     // When:
-    AvroUtil.throwOnInvalidSchemaEvolution(STATEMENT_TEXT, ddlCommand, srClient);
+    AvroUtil.throwOnInvalidSchemaEvolution(STATEMENT_TEXT, ddlCommand, srClient, ksqlConfig);
   }
 
   @Test
@@ -241,7 +269,7 @@ public class AvroUtilTest {
     expectedException.expectMessage("Could not connect to Schema Registry service");
 
     // When:
-    AvroUtil.throwOnInvalidSchemaEvolution(STATEMENT_TEXT, ddlCommand, srClient);
+    AvroUtil.throwOnInvalidSchemaEvolution(STATEMENT_TEXT, ddlCommand, srClient, ksqlConfig);
   }
 
   @Test
@@ -255,7 +283,7 @@ public class AvroUtilTest {
     expectedException.expectMessage("Could not check Schema compatibility");
 
     // When:
-    AvroUtil.throwOnInvalidSchemaEvolution(STATEMENT_TEXT, ddlCommand, srClient);
+    AvroUtil.throwOnInvalidSchemaEvolution(STATEMENT_TEXT, ddlCommand, srClient, ksqlConfig);
   }
 
   private static LogicalSchema toKsqlSchema(final String avroSchemaString) {
